+++ common/src/main/java/com/majruszsaccessories/accessories/AncientScarab.java
@@ -5,6 +5,7 @@ import com.majruszlibrary.data.Reader;
 import com.majruszlibrary.events.OnLootGenerated;
 import com.majruszlibrary.events.base.Condition;
 import com.majruszlibrary.math.Range;
+import com.majruszlibrary.text.RegexString;
 import com.majruszsaccessories.MajruszsAccessories;
 import com.majruszsaccessories.accessories.components.AccessoryIncompatibility;
 import com.majruszsaccessories.accessories.components.BrushingExtraItem;
@@ -14,7 +15,6 @@ import com.majruszsaccessories.common.BonusHandler;
 import com.majruszsaccessories.common.components.TradeOffer;
 import com.majruszsaccessories.events.base.CustomConditions;
 import com.majruszsaccessories.items.AccessoryItem;
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.level.storage.loot.BuiltInLootTables;
 
@@ -34,14 +34,15 @@ public class AncientScarab extends AccessoryHandler {
 
 	static class SuspiciousBlocksDropChance extends BonusComponent< AccessoryItem > {
 		float chance = 0.125f;
-		List< ResourceLocation > locations = List.of(
-			BuiltInLootTables.DESERT_WELL_ARCHAEOLOGY,
-			BuiltInLootTables.DESERT_PYRAMID_ARCHAEOLOGY,
-			BuiltInLootTables.TRAIL_RUINS_ARCHAEOLOGY_COMMON,
-			BuiltInLootTables.TRAIL_RUINS_ARCHAEOLOGY_RARE,
-			BuiltInLootTables.OCEAN_RUIN_WARM_ARCHAEOLOGY,
-			BuiltInLootTables.OCEAN_RUIN_COLD_ARCHAEOLOGY
-		);
+		List< RegexString > locations = RegexString.toRegex( List.of(
+			BuiltInLootTables.DESERT_WELL_ARCHAEOLOGY.toString(),
+			BuiltInLootTables.DESERT_PYRAMID_ARCHAEOLOGY.toString(),
+			BuiltInLootTables.TRAIL_RUINS_ARCHAEOLOGY_COMMON.toString(),
+			BuiltInLootTables.TRAIL_RUINS_ARCHAEOLOGY_RARE.toString(),
+			BuiltInLootTables.OCEAN_RUIN_WARM_ARCHAEOLOGY.toString(),
+			BuiltInLootTables.OCEAN_RUIN_COLD_ARCHAEOLOGY.toString(),
+			"{regex}.*archaeology.*"
+		) );
 
 		public static ISupplier< AccessoryItem > create() {
 			return SuspiciousBlocksDropChance::new;
@@ -54,12 +55,12 @@ public class AncientScarab extends AccessoryHandler {
 				.addCondition( Condition.isLogicalServer() )
 				.addCondition( Condition.predicate( data->data.entity instanceof ServerPlayer ) )
 				.addCondition( CustomConditions.dropChance( ()->this.chance, data->data.entity ) )
-				.addCondition( data->this.locations.contains( data.lootId ) )
+				.addCondition( data->this.locations.stream().anyMatch( id->id.matches( data.lootId.toString() ) ) )
 				.addCondition( data->data.origin != null );
 
 			handler.getConfig()
 				.define( "suspicious_block_spawn_chance", Reader.number(), s->this.chance, ( s, v )->this.chance = Range.CHANCE.clamp( v ) )
-				.define( "suspicious_block_ids", Reader.list( Reader.location() ), s->this.locations, ( s, v )->this.locations = v );
+				.define( "suspicious_block_ids", Reader.list( Reader.string() ), s->RegexString.toString( this.locations ), ( s, v )->this.locations = RegexString.toRegex( v ) );
 		}
 	}
 }